实现生成搜索页分类和品牌过滤菜单

+++ ly-common/src/main/java/com/leyou/item/common/pojo/PageResult.java
@@ -16,7 +16,7 @@ import java.util.List;
 @AllArgsConstructor
 public class PageResult<T> {
     private Long total;// 总条数
-    private Long totalPage;// 总页数
+    private Integer totalPage;// 总页数
     private List<T> items;// 当前页数据
 
     public PageResult(Long total, List<T> items) {
